created a folder components and moved the components into it also change imports
