+++ src/modules/aboutPagePartners/components/AboutPagePartners/AboutPagePartners.module.scss
@@ -1,4 +1,3 @@
-@import '../../../shared/styles/_mixins';
 
 .partnersSection {
   padding: 48px 0;
