+++ src/modules/aboutPagePartners/components/AboutPagePartnersCard/AboutPagePartnersCard.jsx
