+++ src/modules/aboutPagePartners/components/AboutPagePartnersCard/AboutPagePartnersCard.module.scss
@@ -1,4 +1,3 @@
-@import '../../../shared/styles/_mixins';
 
 .imageContainer {
   width: 162px;
