+++ src/modules/aboutPagePartners/components/AboutPagePartnersCarousel/AboutPagePartnersCarousel.module.scss
@@ -1,4 +1,3 @@
-@import '../../../shared/styles/_mixins';
 
 .list {
   display: flex;
